+++ src/CfdiUtils/Elements/CartaPorte30/DocumentacionAduanera.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CfdiUtils\Elements\CartaPorte30;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class DocumentacionAduanera extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'cartaporte30:DocumentacionAduanera';
+    }
+}